feat: add file read functionality to getHubVersionInformation
This commit introduces the use of 'fs' and 'path' from Node.js in the 'getHubVersionInformation' function on the AppController. The function will now read and return data from a file in the 'updateData' folder matching the provided version number. A 'POST' request is now used instead of a 'GET' request to retrieve this