refactor: replace ES6 import with require for Agenda module
This commit replaces the ES6 import statement for the Agenda module with a require statement to prevent compatibility issues. Also, the type of 'agenda' variable has been changed from 'Agenda' to 'any' to allow for more flexibility in the type of values it can hold. This change does not affect the functionality of the code.