feat: Added sectors to ATSUInformation
This commit includes changes in the app.controller to include 'sectors' in the ATSUInformation. This will provide more specific details about the user's station. Additionally, 'acars_user_id' was replaced with 'cid' for better semantics and clarity. Updates were made to ensure that changes in sectors are reflected in the ATSUInformation.