feat: Update app.service to reflect new plugin version
This commit includes changes in app.service file to reflect the new version of vatACARS plugin for vatSys. The plugin is now in 'Early Access' with version '1.0.1'. The status has been updated to 'waiting' and the local version has been cleared. The version update also includes the initial release of the vatACARS plugin for vatSys in the changes array.