feat: added dark theme to swagger documentation
This commit includes the addition of a dark theme to the swagger documentation. This was achieved by introducing the 'swagger-themes' package to the project. The theme 'DARK_MONOKAI' was selected and applied to the swagger documentation. This feature enhances the user experience by providing a darker, less strain inducing option for viewing the documentation.